refactor: result page
