+++ src/pages/result/FinishedMissionList.tsx
