+++ src/pages/result/OverallStatus/BannerSection.tsx
