+++ src/pages/result/OverallStatus/ImageAreaSection.tsx
