+++ src/pages/result/OverallStatus/MissionSection.tsx
