+++ src/pages/result/result.constants.ts
